Refactor datastore.
