+++ package.json
@@ -259,7 +259,7 @@
     "jquery": "^1.12.4",
     "metrics-graphics": "^2.15.6",
     "node-sass": "^4.9.2",
-    "npm": "^6.1.0",
+    "npm": "^6.4.1",
     "plotly.js": "^1.40.1",
     "react": "^16.0.0",
     "react-bootstrap": "^0.32.1",
