+++ src/app.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-
+import {DataStore} from "./datastore.jsx";
 import "./app.css";
 
 import {Navigation} from "./components/navigation.jsx";
@@ -8,13 +8,10 @@ import {MetricSelector} from "./components/metricselector.jsx";
 import {VersionSelector} from "./components/versionselector.jsx";
 import {ChannelSelector} from "./components/channelselector.jsx";
 
-import {MetricData} from "./metricdata.js";
-
 class App extends React.Component {
   constructor(props) {
     super(props);
-
-    this.dataStore = new MetricData();
+    this.dataStore = new DataStore();
     let active = this.dataStore.active;
     this.state = {
       activeMetric: active.metric,
